Fix outlined textfield notch when initially rendered hidden (#488)
fixes #444 by providing the mwc-textfied.layout() method.

In this PR:

added layout function
notch performs layout when label is changed
fixed hidden bug where foundation was being recreated too often
caused bug where changing maxLength caused foundation to unset invalid
had to pull logic out of foundation and into registerValidationAttributeHandler
The issue: when the label is floating and the element is outlined, but the label has no width, e.g.
mwc-textfield[outlined][label="something"][value="something"][hidden]

the mdc dataflow will calculate the size of the notch as zero. When hidden is removed, the notch will not update in size. Here we give them the layout method.

It is not possible to know when the textfield is changed from display: none to visible without setting a resize observer on the label. I wanted to avoid that because of possible render trashing and the overall heaviness of resizeObserver and a polyfill. Instead, it is up to the user to call layout when they change the visibility of the element.